daemon/requirements.txt
greenlet==2.0.2
PyMySQL==1.1.0
python-dotenv==1.0.0
SQLAlchemy==2.0.19
typing_extensions==4.7.1
